Remove call to logging.basicConfig()
logging.basicConfig() is something which can only be done once
per process.

Simply importing a library shouldn't have the side-effect of
calling this. It means for example that a command-line tool's
call to logging.basicConfig with some arguments will be ignored;
and if multiple libraries do this with different arguments, it
means log behavior mysteriously changes depending on the order
in which modules are imported.