Minor refactor & raise coverage to 100%
When YAML can't be parsed, the library logs a message telling which
file it was attempting to load when it found invalid YAML.

For some reason, that log was only in the remote code path.
There doesn't seem to be any particular reason for that; why wouldn't
you want the same message logged for both remote and local cases?
Refactor to do this & make it tested.