Refactor get_loader, Loader concepts
The new interface is simply:

- get_loader(source=None), where source is a local path or URL
- if source is omitted then use DEFAULT_UBI_REPO

In comparison the old interface of get_loader(local, local_repo)
was messy with a few problems:

---

It had three different "modes": remote repo, local repo, or just
local (meaning absolute paths must be given).

Supporting both the "local repo" and "local" modes don't
seem to make much sense to me.  With these two modes, you have to
know whether the loader was created in "local repo" mode or "local"
mode before you know how to use it (because you have to know whether
you're supposed to give relative or absolute paths). So it's hard
to write a function which can take a loader and work regardless of
which data source is used behind the loader.  Let's collapse this
into having only two modes, local and remote, both of which load
files using relative paths.

---

The implementation of all three modes was bundled into
a single Loader class.  The main methods in that class had
different code paths for the different modes, and some methods
only called in certain modes.  This is a natural fit for
implementing in separate classes instead of having the code
together unnecessarily, so let's do that.

---

It was not possible to have DEFAULT_UBI_REPO set to anything other
than a gitlab repo; and it was not possible to use a gitlab repo
by any means other than setting DEFAULT_UBI_REPO. There doesn't
seem to be any reason for such restrictions. It means for instance
you couldn't write code which would normally access a remote repo,
and simply redirect it to a local directory by setting
DEFAULT_UBI_REPO.

Let's just let the caller pass what they want, and use
DEFAULT_UBI_REPO if they haven't passed anything.

---

Lots of internal-only attributes like repo_api were available on
the Loader, with no hints these are not meant for public use.
Make sure they're named with leading _ so it's clear they're
private.