Generic GitLab loader [RHELDST-19139]
Added support for generic (custom) names of branches
for GitlabLoader of ubi-config.

Removed also some hardcoded parts of code that were not
used in practice and were likely broken.

This change makes the `version` argument mandatory for `GitlabLoader.load()`
method even if it has currently default value set to None
(keeping the signature of overridden method). We can't easily guess what's
the default (fallback) branch in repository with multiple branches for each major version.

In practice we mostly rely on `load_all()` method that always provides
`version` arg in GitlabLoader.

This generic improvement is not supported for LocalLoader which is not
currently required for generic (non-ubi) config files.