Support for flags [RHELDST-18810]
Added support for flags to be present in ubi-config files.
They can be defined as a dictionary under the `flags` field in
the config file.

These flags can be used in various tools that use ubi-config library.

This commit adds support for flag `base_pkgs_only` that will tell
ubi-manifest service to include only the pkgs that are defined
in the config file and nothing else.